+++ agent/custom/reco/__init__.py
@@ -1,9 +1,5 @@
 from .climb_tower_quiz import *
-from .climb_tower_potential import *
-from .climb_tower_potential_bag import *
 
 __all__ = [
-    "QuizRecognition",
-    "ChoosePotentialRecognition",
-    "PotentialBagTestRecognition"
+    "QuizRecognition"
 ]
