--- agent/custom/reco/climb_tower_potential.py
@@ -1,1543 +1,0 @@
-from __future__ import annotations
-
-import re
-import time
-from dataclasses import dataclass, field, replace
-from typing import Optional, Any, Self
-from math import comb
-
-import numpy
-
-from maa.agent.agent_server import AgentServer
-from maa.custom_recognition import CustomRecognition
-from maa.context import Context
-
-from utils import logger as logger_module
-logger = logger_module.get_logger("climb_tower_potential")
-
-
-# region 常量 ===============================================================
-
-MAX_POTENTIAL_LEVEL: int = 6  # 潜能等级上限，condition max_level 字段的默认值
-
-
-# endregion 常量 ===============================================================
-
-# region 跨节点储存的状态类 ===============================================================
-
-@dataclass(slots=True)
-class OwnedPotential:
-    name: str
-    level: int
-    recommended_level: int
-    trekker: str = "unknown"
-    core: bool = False
-
-    @property
-    def max_level(self) -> int:
-        """潜能等级上限"""
-        return 1 if self.core else MAX_POTENTIAL_LEVEL
-
-    @property
-    def owned(self) -> bool:
-        """是否已持有"""
-        return self.level > 0
-
-    @property
-    def recommended(self) -> bool:
-        """是否为推荐潜能"""
-        return self.recommended_level > 0
-
-@dataclass(slots=True)
-class OwnedPotentials:
-    potentials: list[OwnedPotential] = field(default_factory=list)
-
-    def __iter__(self):
-        return iter(self.potentials)
-
-    def __len__(self):
-        return len(self.potentials)
-
-    def save(self, potential: Potential, *, fuzzy: bool = False) -> None:
-        """将选中的潜能保存到OwnedPotentials中，如果已存在则更新等级和名字"""
-        if not potential.name:
-            return
-
-        trekker = potential.trekker or "unknown"
-        level = max(potential.new_level, 1)
-        core = potential.core
-
-        # 先查找潜能是否已存在，如果存在则更新等级和名字
-        existed = self.find(potential.name, trekker=trekker, core=core, fuzzy=fuzzy)
-        if existed:
-            # 防止 OCR 识别失败导致等级倒退，所以至少按照原等级+1处理
-            existed.level = min(existed.max_level, max(existed.level + 1, level))
-            existed.recommended_level = max(existed.recommended_level, potential.recommended_level)
-            # 使用更长的名字，因为更长的名字更接近原名
-            existed.name = self._longer_name(existed.name, potential.name)
-            return
-
-        # 不存在，就添加到list中
-        self.potentials.append(OwnedPotential(
-            name=potential.name,
-            level=level,
-            recommended_level=potential.recommended_level,
-            trekker=trekker,
-            core=core,
-        ))
-
-    def find(
-        self,
-        name: str,
-        *,
-        trekker: str | None = None,
-        core: bool | None = None,
-        fuzzy: bool = False,
-    ) -> OwnedPotential | None:
-        """查找是否已拥有该潜能"""
-        # TODO: 增加difflib匹配
-        for p in self.potentials:
-            if trekker is not None and p.trekker != trekker:
-                continue
-            if core is not None and p.core != core:
-                continue
-            if self._match(p.name, name, fuzzy=fuzzy):
-                return p
-        return None
-
-    def find_level(
-        self,
-        name: str,
-        *,
-        trekker: str | None = None,
-        fuzzy: bool = False,
-    ) -> int:
-        """查找已拥有的某潜能的等级，如果不存在则返回0"""
-        owned_potential = self.find(name, trekker=trekker, fuzzy=fuzzy)
-        return owned_potential.level if owned_potential else 0
-
-    def find_recommended_level(
-        self,
-        name: str,
-        *,
-        trekker: str | None = None,
-        fuzzy: bool = False,
-    ) -> int:
-        """查找已拥有的某潜能的推荐等级，如果不存在则返回0"""
-        owned_potential = self.find(name, trekker=trekker, fuzzy=fuzzy)
-        return owned_potential.recommended_level if owned_potential else 0
-
-    def count(
-        self,
-        *,
-        trekker: str | list = None,
-        level_at_least: int | None = None,
-        level_at_most: int | None = None,
-        recommended_level_at_least: int | None = None,
-        recommended_level_at_most: int | None = None,
-        include_core: bool = False,
-        incomplete_only: bool = False,
-        leveling_only: bool = False
-    ) -> int:
-        """统计符合条件的潜能的数量"""
-        if isinstance(trekker, str):
-            trekker = (trekker,)
-
-        return sum(
-            1 for potential in self.potentials
-            if (trekker is None or potential.trekker in trekker)
-            and (level_at_least is None or potential.level >= level_at_least)
-            and (level_at_most is None or potential.level <= level_at_most)
-            and (recommended_level_at_least is None or potential.recommended_level >= recommended_level_at_least)
-            and (recommended_level_at_most is None or potential.recommended_level <= recommended_level_at_most)
-            and (include_core or not potential.core)
-            and (not incomplete_only or potential.level < potential.recommended_level)
-            and (not leveling_only or potential.level < potential.max_level)
-        )
-
-    @staticmethod
-    def _match(a: str, b: str, *, fuzzy: bool) -> bool:
-        """
-        判断两个字符串是否匹配。
-        Args:
-            a: 第一个字符串
-            b: 第二个字符串
-            fuzzy: 是否开启模糊匹配（通过in方法匹配，只能处理前后有漏的情况）
-
-        Returns:
-            bool: 是否匹配成功
-        """
-        # 处理特殊符号
-        a = re.sub(r"\W", "", a)
-        b = re.sub(r"\W", "", b)
-
-        # 精确匹配
-        if not fuzzy:
-            return a == b
-
-        # 简单处理日语的长音符号，中文的一不处理是因为会出现重名
-        a = a.replace("ー", "")
-        b = b.replace("ー", "")
-
-        # 通过in进行模糊匹配，需要排除空字符串然后再判断包含关系
-        return bool(a and b and (a in b or b in a))
-
-    @staticmethod
-    def _longer_name(old: str, new: str) -> str:
-        """一般来讲，更长的名称更接近原名"""
-        return new if len(new) > len(old) else old
-
-
-class State:
-    failed_count: int = 0
-    high_level_span_count: int = 0
-    enhance_high_level_span_count: int = 0
-    potentials_level_count: int = 0
-    main_trekker: str = ""
-    trekker_images: list[numpy.ndarray] = []
-    owned_potentials: OwnedPotentials = OwnedPotentials()
-
-    @classmethod
-    def reset(cls):
-        cls.failed_count = 0
-        cls.high_level_span_count = 0
-        cls.enhance_high_level_span_count = 0
-        cls.potentials_level_count = 0
-        cls.main_trekker = ""
-        cls.trekker_images.clear()
-        cls.owned_potentials = OwnedPotentials()
-
-# endregion 跨节点储存的状态类 ===============================================================
-
-# region 潜能选择界面布局类 ===============================================================
-
-DEFAULT_POTENTIAL_LAYOUTS = {
-    1: [
-        {
-            "core_potential_roi": [530, 405, 220, 60],
-            "general_potential_roi": [530, 375, 220, 60],
-            "general_potential_level_roi": [530, 425, 220, 40],
-            "recommended_level_roi": [670, 165, 140, 50],
-            "potential_roi": [470, 0, 343, 720],
-            "trekker_roi": [500, 182, 40, 40],
-            "x_border": [470, 813]
-        }
-    ],
-    2: [
-        {
-            "core_potential_roi": [358, 405, 220, 60],
-            "general_potential_roi": [358, 375, 220, 60],
-            "general_potential_level_roi": [358, 425, 220, 40],
-            "recommended_level_roi": [490, 165, 140, 50],
-            "potential_roi": [0, 0, 639, 720],
-            "trekker_roi": [329, 182, 40, 40],
-            "x_border": [0, 639]
-        },
-        {
-            "core_potential_roi": [703, 405, 220, 60],
-            "general_potential_roi": [703, 375, 220, 60],
-            "general_potential_level_roi": [703, 425, 220, 40],
-            "recommended_level_roi": [840, 165, 140, 50],
-            "potential_roi": [640, 0, 640, 720],
-            "trekker_roi": [673, 182, 40, 40],
-            "x_border": [640, 1280]
-        }
-    ],
-    3: [
-        {
-            "core_potential_roi": [187, 405, 220, 60],
-            "general_potential_roi": [187, 375, 220, 60],
-            "general_potential_level_roi": [187, 425, 220, 40],
-            "recommended_level_roi":[320, 165, 140, 50],
-            "potential_roi": [0, 0, 469, 720],
-            "trekker_roi": [156, 182, 40, 40],
-            "x_border": [0, 469]
-        },
-        {
-            "core_potential_roi": [530, 405, 220, 60],
-            "general_potential_roi": [530, 375, 220, 60],
-            "general_potential_level_roi": [530, 425, 220, 40],
-            "recommended_level_roi": [670, 165, 140, 50],
-            "potential_roi": [470, 0, 343, 720],
-            "trekker_roi": [500, 182, 40, 40],
-            "x_border": [470, 813]
-        },
-        {
-            "core_potential_roi": [875, 405, 220, 60],
-            "general_potential_roi": [875, 375, 220, 60],
-            "general_potential_level_roi": [875, 425, 220, 40],
-            "recommended_level_roi":[1010, 165, 140, 50],
-            "potential_roi": [814, 0, 466, 720],
-            "trekker_roi": [844, 182, 40, 40],
-            "x_border": [814, 1280]
-        }
-    ]
-}
-
-@dataclass(slots=True)
-class PotentialLayout:
-    core_potential_roi: list[int]
-    general_potential_roi: list[int]
-    general_potential_level_roi: list[int]
-    recommended_level_roi: list[int]
-    potential_roi: list[int]
-    trekker_roi: list[int]
-    x_border: list[int]
-
-@dataclass(slots=True)
-class PotentialLayouts:
-    potential_layouts: dict[int, list[PotentialLayout]] = field(default_factory=lambda: {
-        count: [PotentialLayout(**l) for l in layouts]
-        for count, layouts in DEFAULT_POTENTIAL_LAYOUTS.items()
-    })
-
-    def __getitem__(self, key):
-        return self.potential_layouts[key]
-
-    def __iter__(self):
-        return iter(self.potential_layouts)
-
-    def items(self):
-        return self.potential_layouts.items()
-
-    def get(self, key, default=None):
-        return self.potential_layouts.get(key, default)
-
-
-# endregion 潜能选择界面布局类 ===============================================================
-
-# region 节点参数类 ===============================================================
-
-@dataclass(slots=True)
-class Parameters:
-    trigger_type: str
-    max_refresh_count: int
-    reserved_coin: int
-    priority_list: list[dict]
-    handler: str
-    chooser: str
-    tower_8_strategy: str
-    tower_8_capacity_strategy: str
-    # 自带默认值的参数
-    potential_layouts: PotentialLayouts = field(default_factory=lambda: PotentialLayouts())
-    selected_potential_offset: int = 35
-
-    def __post_init__(self):
-        parsed_list = self._parse_priority_raw_list(
-            self.priority_list,
-            State.owned_potentials,
-        )
-        object.__setattr__(self, 'priority_list', parsed_list)
-
-    @staticmethod
-    def _parse_priority_raw_list(
-        potential_priority_raw: list[dict],
-        owned_potentials: OwnedPotentials,
-    ) -> list[dict[str, Any]]:
-        """对原始 priority_list 进行初筛，过滤掉 condition 当前不满足的规则。
-
-        排名使用原始 JSON 的 1-based 行号（index + 1），数值越小排名越高。
-        condition 不满足的条目直接跳过，其行号仍保留在原始位置，不影响其他条目的排名。
-
-        Args:
-            potential_priority_raw: 原始优先级列表，每个元素结构：
-                {
-                    "trekker": str,         # 可选，潜能归属角色名
-                    "potential": str|list,  # 必填，潜能名称或名称列表
-                    "level_span": int,      # 可选，默认 1，最小升级跨度
-                    "max_level": int,       # 可选，默认 MAX_POTENTIAL_LEVEL，旧等级上限（不含）
-                    "refresh": int,         # 可选，默认 0，已刷新次数必须 >= 该值规则才生效
-                    "condition": list       # 可选，生效条件，元素为 dict 时 AND，为 list 时 OR
-                }
-            owned_potentials: 已拥有潜能状态。
-
-        Returns:
-            list[dict]: 通过初筛的规则列表，每个元素结构：
-                {
-                    "trekker": str | None,  # 归属角色
-                    "names": list[str],     # 目标潜能名称列表
-                    "level_span": int,      # 最小升级跨度
-                    "max_level": int,       # 旧等级上限（不含）
-                    "refresh": int,         # 已刷新次数下限
-                    "priority": int         # 原始 JSON 的 1-based 行号，越小排名越高
-                }
-        """
-        def _check_single_condition(item: dict) -> bool:
-            """检查单个 condition 子项是否满足。"""
-            if "count_at_least" in item or "count_at_most" in item:
-                level_min = item.get("level_at_least")
-                level_max = item.get("level_at_most")
-                count = owned_potentials.count(
-                    trekker=item["trekker"],
-                    level_at_least=level_min,
-                    level_at_most=level_max,
-                )
-                if "count_at_least" in item and count < item["count_at_least"]:
-                    return False
-                if "count_at_most" in item and count > item["count_at_most"]:
-                    return False
-                return True
-            current = owned_potentials.find_level(
-                item["potential"],
-                trekker=item.get("trekker"),
-            )
-            min_ok = current >= item["level_at_least"] if "level_at_least" in item else True
-            max_ok = current <= item["level_at_most"] if "level_at_most" in item else True
-            return min_ok and max_ok
-
-        def _check_condition(cond: list) -> bool:
-            """检查 condition 列表是否满足。
-
-            元素全为 dict 时为 AND 逻辑；含 list 元素时为 OR 逻辑（内层为 AND）。
-            """
-            if not cond:
-                return True
-            if all(isinstance(item, dict) for item in cond):
-                return all(_check_single_condition(item) for item in cond)
-            for branch in cond:
-                if isinstance(branch, list):
-                    if all(_check_single_condition(item) for item in branch):
-                        return True
-                elif isinstance(branch, dict):
-                    if _check_single_condition(branch):
-                        return True
-            return False
-
-        valid_entries = []
-        for index, raw in enumerate(potential_priority_raw):
-            condition = raw.get("condition", [])
-            if not isinstance(condition, list):
-                continue
-            if not _check_condition(condition):
-                continue
-
-            potential = raw["potential"]
-            names = potential if isinstance(potential, list) else [potential]
-
-            valid_entries.append({
-                "trekker": raw.get("trekker"),
-                "names": names,
-                "level_span": raw.get("level_span", 1),
-                "max_level": raw.get("max_level", MAX_POTENTIAL_LEVEL),
-                "refresh": raw.get("refresh", 0),
-                "priority": index + 1,
-            })
-
-        return valid_entries
-
-
-# endregion 节点参数类 ===============================================================
-
-# region 统筹所有数据的类（包括节点参数） ===============================================================
-
-@dataclass(slots=True)
-class Data:
-    params: Parameters
-    current_coin: int = 0
-    refresh_cost: int = 0
-    potential_count: int = 0
-    core_potential: bool = False
-    # 需要根据刷新更新的数据
-    selected_potential_index: int = 1
-    potentials: list[Potential] = field(default_factory=lambda: [])
-    # 计数用数据
-    refresh_count: int = 0
-
-    @property
-    def refresh_botton(self) -> bool:
-        return self.refresh_cost >= 0
-
-    @property
-    def refreshable(self) -> bool:
-        return self.refresh_count < self.refresh_limit
-
-    @property
-    def refresh_limit(self) -> int:
-        usable_coin = max(0, self.current_coin - self.params.reserved_coin)
-        affordable = usable_coin // self.refresh_cost if self.refresh_botton else 0
-        return min(self.params.max_refresh_count, affordable)
-
-    @property
-    def potential_rois(self) -> list[list[int]]:
-        return [l.potential_roi for l in self.params.potential_layouts[self.potential_count]]
-
-    @property
-    def x_borders(self) -> list[list[int]]:
-        return [l.x_border for l in self.params.potential_layouts[self.potential_count]]
-
-    @property
-    def core_potential_name_rois(self) -> list[list[int]]:
-        return [l.core_potential_roi for l in self.params.potential_layouts[self.potential_count]]
-
-    @property
-    def general_potential_name_rois(self) -> list[list[int]]:
-        return [l.general_potential_roi for l in self.params.potential_layouts[self.potential_count]]
-
-    @property
-    def general_potential_level_rois(self) -> list[list[int]]:
-        return [l.general_potential_level_roi for l in self.params.potential_layouts[self.potential_count]]
-
-    @property
-    def recommended_level_rois(self) -> list[list[int]]:
-        return [l.recommended_level_roi for l in self.params.potential_layouts[self.potential_count]]
-
-    @property
-    def trekker_rois(self) -> list[list[int]]:
-        return [l.trekker_roi for l in self.params.potential_layouts[self.potential_count]]
-
-@dataclass(slots=True)
-class Potential:
-    layout: PotentialLayout
-    index: int = -1
-    core: bool = False
-    name: str = ""
-    old_level: int = -1
-    new_level: int = -1
-    recommended: bool = False
-    recommended_level: int = -1
-    trekker: str = ""
-    selected: bool = False
-    # 自定义参数
-    rank: int = -1
-    sub_rank: int = -1
-    score: int = 0
-
-    @property
-    def level_span(self) -> int:
-        return self.new_level - self.old_level
-
-    @property
-    def box(self) -> list[int]:
-        return self.layout.general_potential_roi
-
-    @property
-    def potential_roi(self) -> list[int]:
-        return self.layout.potential_roi
-
-    @property
-    def x_border(self) -> list[int]:
-        return self.layout.x_border
-
-    @property
-    def core_potential_name_roi(self) -> list[int]:
-        return self.layout.core_potential_roi
-
-    @property
-    def general_potential_name_roi(self) -> list[int]:
-        return self.layout.general_potential_roi
-
-    @property
-    def general_potential_level_roi(self) -> list[int]:
-        return self.layout.general_potential_level_roi
-
-    @property
-    def recommended_level_roi(self) -> list[int]:
-        return self.layout.recommended_level_roi
-
-    def update(self, screen: ScreenDataProcessor, data: Data):
-        # 更新核心潜能
-        if data.core_potential:
-            self.core = True
-
-        # 更新潜能数据
-        self.name = self._get_name(screen, data)
-        self.old_level, self.new_level = self._get_level(screen, data)
-        self.recommended, self.recommended_level = self._get_recommended_data(screen, data)
-
-    def _get_name(self, screen: ScreenDataProcessor, data: Data) -> str:
-        roi = self.core_potential_name_roi if self.core else self.general_potential_name_roi
-        adjusted_roi = self._get_adjusted_roi(roi, data.params.selected_potential_offset)
-        return screen.get_potential_name(adjusted_roi)
-
-    def _get_level(self, screen: ScreenDataProcessor, data: Data) -> tuple[int, int]:
-        if self.core:
-            return 0, 1
-        adjusted_roi = self._get_adjusted_roi(self.general_potential_level_roi, data.params.selected_potential_offset)
-        old, new = screen.get_potential_level(adjusted_roi)
-        return old, new
-
-    def _get_recommended_data(self, screen: ScreenDataProcessor, data: Data) -> tuple[bool, int]:
-        roi = self.potential_roi
-        adjusted_roi = [self._get_adjusted_roi(roi, data.params.selected_potential_offset)]
-        recommended = True if screen.check_potential_recommended(adjusted_roi) else False
-        if not recommended:
-            return False, 0
-        if recommended and data.core_potential:
-            return True, 1
-        roi = self.recommended_level_roi
-        adjusted_roi = self._get_adjusted_roi(roi, data.params.selected_potential_offset)
-        level = screen.get_recommend_level(adjusted_roi)
-        return recommended, level
-
-    def _get_adjusted_roi(self, roi, offset) -> list[int]:
-            return [roi[0], max(0, roi[1] - offset) if self.selected else roi[1], roi[2], roi[3]]
-
-
-
-# endregion 统筹所有数据的类（包括节点参数） ===============================================================
-
-# region 使用maa与屏幕交互类 ===============================================================
-
-class ScreenDataProcessor:
-    def __init__(self, context: Context):
-        self.context = context
-        self.image = None
-        self.max_try = 1
-
-    def load_last_screenshot(self):
-        self.image = self.context.tasker.controller.cached_image
-
-    def screenshot(self):
-        self.image = self.context.tasker.controller.post_screencap().wait().get()
-
-    def crop_screenshot(self, roi: list[int]) -> numpy.ndarray:
-        return self.image[roi[1]:roi[1] + roi[3], roi[0]:roi[0] + roi[2]]
-
-    def refresh(self):
-        self.context.run_task("星塔_节点_选择潜能_点击刷新_agent")
-
-    def click_potential(self, box: list[int]) -> bool:
-        """点击指定box"""
-        pipeline_override = {
-            "星塔_节点_选择潜能_点击潜能_agent": {
-                "action": {
-                    "param": {
-                        "target": box
-                    }
-                }
-            }
-        }
-        run_result = self.context.run_task("星塔_节点_选择潜能_点击潜能_agent", pipeline_override)
-        if run_result and run_result.status.succeeded:
-            return True
-        return False
-
-    def _base_recognition(
-            self, mode, node_name, failed_return, *, roi=None, image=None, template=None, max_try=0
-    ) -> Any:
-        """
-        核心识别逻辑：支持 OCR 和 Template
-        识别失败时返回默认值
-
-        Args:
-            mode(str): 识别模式，"ocr"或"template"或"color"
-            node_name(str): 节点名称，用于识别结果的记录和返回
-            failed_return(any): 识别失败时的默认值
-            roi(tuple): 可选的ROI坐标，用于模板识别
-            image(numpy.ndarray): 可选的自定义图像，用于识别
-            template(str): 可选的模板名称，用于模板识别
-            max_try(int): 可选的最大重试次数，默认为1
-
-        Returns:
-            any: 识别到的结果，根据mode返回文本或坐标列表
-        """
-        if image is None:
-            if self.image is None:
-                self.image = self.context.tasker.controller.post_screencap().wait().get()
-            image = self.image
-
-        actual_max_try = max_try if max_try > 0 else self.max_try
-
-        params = {}
-        if roi:
-            params["roi"] = roi
-        if template and mode == "template":
-            params["template"] = template
-        pipeline_override = {node_name: {"recognition": {"param": params}}} if params else {}
-
-        try_count = 0
-        while True:
-            reco_detail = self.context.run_recognition(node_name, image, pipeline_override)
-
-            if reco_detail and reco_detail.hit:
-                if mode == "ocr":
-                    # OCR 逻辑：返回文本列表，包含坐标和分数
-                    logger.debug(
-                        f"节点{node_name} OCR结果：{[(r.text, r.box, r.score) for r in reco_detail.filtered_results]}"
-                    )
-                    results = reco_detail.filtered_results
-                    return [(r.text, r.box) for r in results]
-                elif mode == "color":
-                    # Color 逻辑：返回布尔结果
-                    return True if reco_detail.filtered_results else False
-                else:
-                    # Template 逻辑：返回坐标列表，包含分数
-                    logger.debug(f"节点{node_name} 模板结果：{[(r.box, r.score) for r in reco_detail.filtered_results]}")
-                    results = sorted(reco_detail.filtered_results, key=lambda r: r.score, reverse=True)
-                    return [r.box for r in results]
-
-            # 统一的日志记录
-            status = "未识别到有效结果" if reco_detail and reco_detail.all_results else "未识别到任何内容"
-            logger.debug(f"节点'{node_name}'{status}")
-
-            if self.context.tasker.stopping:
-                return failed_return
-
-            try_count += 1
-            if try_count >= actual_max_try:
-                break
-
-            logger.debug("等待1秒后重新识别")
-            time.sleep(1)
-            image = self.context.tasker.controller.post_screencap().wait().get()
-
-        logger.debug(f"无法识别节点'{node_name}'，返回默认值 {failed_return}")
-        return failed_return
-
-    def _ocr(self, node_name, failed_return, **kwargs) -> list[tuple[str, list[int]]]:
-        return self._base_recognition("ocr", node_name, failed_return, **kwargs)
-
-    def _template(self, node_name, failed_return, **kwargs) -> list[list[int]]:
-        return self._base_recognition("template", node_name, failed_return, **kwargs)
-
-    def _color(self, node_name, **kwargs) -> bool:
-        return self._base_recognition("color", node_name, False, **kwargs)
-
-    def get_current_coin(self, image: Optional[numpy.ndarray] = None, max_try: int = 1) -> int:
-        ocr_results = self._ocr("星塔_通用_识别当前金币_agent", [["0"]], image=image, max_try=max_try)
-        return int(ocr_results[0][0])
-
-    def get_refresh_cost(self, image: Optional[numpy.ndarray] = None, max_try: int = 1) -> int:
-        ocr_results = self._ocr("星塔_通用_识别刷新花费_agent", [["-1"]], image=image, max_try=max_try)
-        return int(ocr_results[0][0])
-
-    def check_core_potential(self, image: Optional[numpy.ndarray] = None, max_try: int = 1) -> bool:
-        if self._template("星塔_节点_选择潜能_识别核心潜能_agent", [], image=image, max_try=max_try):
-            return True
-        return False
-
-    def get_potential_name(self, roi: list[int], image: Optional[numpy.ndarray] = None, max_try: int = 1) -> str:
-        node_name = "星塔_节点_选择潜能_识别潜能名称_agent"
-        ocr_results = self._ocr(node_name, [["", []]], roi=roi, image=image, max_try=max_try)
-        return " ".join([t for t, _ in ocr_results])
-
-    def get_potential_level(
-            self,
-            roi: list[int],
-            image: Optional[numpy.ndarray] = None,
-            max_try: int = 1
-    ) -> tuple[int, int]:
-        node_name = "星塔_节点_选择潜能_识别潜能等级_agent"
-        ocr_results = self._ocr(node_name, [["", []]], roi=roi, image=image, max_try=max_try)
-        levels = self._parse_level_text([t for t, _ in ocr_results])
-        return levels
-
-    @staticmethod
-    def _parse_level_text(texts: list[str]) -> tuple[int, int]:
-        """解析 OCR 返回的等级数字结果集。
-
-        pipeline OCR 使用 \\d+ 匹配并剔除语言关键词，可能返回：
-            ["1"]       -> old=0, new=1  （新获得，只有新等级）
-            ["4", "5"]  -> old=4, new=5
-            ["45"]      -> old=4, new=5  （两位数粘连）
-        仅在游戏版本保持最大潜能等级小于10时有效。
-
-        Args:
-            texts: OCR filtered_results 中各结果的 text 列表
-
-        Returns:
-            tuple[int, int]: (old_level, new_level)，解析失败返回 (0, 0)
-        """
-        # 将 ["4", "5"] 或 ["45"] 统一转为 "45"
-        full_text = "".join(t for t in texts if t.isdigit())
-
-        if len(full_text) == 1:
-            return 0, int(full_text)
-        if len(full_text) >= 2:
-            # 取前两个数字处理粘连
-            return int(full_text[0]), int(full_text[1])
-        logger.warning(f"无法解析潜能等级（识别到的等级文本: {texts}）")
-        return -1, -1
-
-    def get_recommend_level(self, roi: list[int], image: Optional[numpy.ndarray] = None, max_try: int = 1) -> int:
-        node_name = "星塔_节点_选择潜能_识别推荐等级_agent"
-        ocr_results = self._ocr(node_name, [["0"]], roi=roi, image=image, max_try=max_try)
-        return int(ocr_results[0][0])
-
-    def check_item_list_visibility(self, max_try: int = 1) -> bool:
-        ocr_results = self._ocr("星塔_节点_选择潜能_检测干扰文字_agent", [], max_try=max_try)
-        return len(ocr_results) != 0
-
-    def get_potential_count(
-            self,
-            core_potential: bool = False,
-            image: Optional[numpy.ndarray] = None,
-            max_try: int = 1
-    ) -> int:
-        """
-            检查可选潜能卡片数量
-
-            Args:
-                core_potential(bool): 是否为核心潜能，默认为False
-                image(nd.array): 截图
-                max_try(int): 可选的最大重试次数，默认为1
-
-            Returns:
-                int: 可选潜能卡片数量，识别失败时返回3
-        """
-        if core_potential:
-            return 3
-
-        node_name = "星塔_节点_选择潜能_识别潜能数量_agent"
-        failed_return = [1, 2, 3]
-        results = self._template(node_name, failed_return, image=image, max_try=max_try)
-        if results == failed_return:
-            logger.error("潜能数量识别失败，将默认为3个潜能")
-        return len(results)
-
-    def get_recommended_potential(
-            self,
-            borders: list[list],
-            image: Optional[numpy.ndarray] = None,
-            max_try: int = 1
-    ) -> list:
-        """
-        识别系统推荐图标，返回对应卡片的潜能序数列表。
-
-        推荐图标位于卡片 box 范围内，通过判断图标命中 x 坐标是否落入各卡片
-        x_border 区间来确定归属卡片。
-        识别失败时返回第一张卡片的 box 作为兜底。
-
-        Args:
-            borders: 可选潜能卡片区域列表，每个元素结构：[float, float],  # 卡片 x 轴边界（左闭右闭）
-            image: 截图
-            max_try: 可选的最大重试次数，默认为1
-
-        Returns:
-            list: 包含推荐潜能序数的列表
-        """
-        recommended_boxes = self._template(
-            "星塔_节点_选择潜能_识别推荐图标_agent",
-            [],
-            image=image,
-            max_try=max_try
-        )
-        hit_xs = [r[0] for r in recommended_boxes]
-        matched = [
-            i for x in hit_xs
-            for i, (low, high) in enumerate(borders)
-            if low <= x <= high
-        ]
-
-        if not matched:
-            logger.debug("推荐图标识别失败，有可能是没有推荐图标，也有可能是识别问题")
-        if len(matched) != len(hit_xs):
-            logger.error("推荐图标识别位置与潜能数量不匹配，潜能选择可能会出现问题")
-
-        return matched
-
-    def check_potential_recommended(
-            self,
-            roi: list[list],
-            image: Optional[numpy.ndarray] = None,
-            max_try: int = 1
-    ) -> bool:
-        node_name = "星塔_节点_选择潜能_识别推荐图标_agent"
-        recommended_boxes = self._template(node_name, [], roi=roi, image=image, max_try=max_try)
-        if recommended_boxes:
-            return True
-        return False
-
-    def get_selected_potential_index(
-            self,
-            borders: list[list[int]],
-            image: Optional[numpy.ndarray] = None,
-            max_try: int = 1
-    ) -> int:
-        """识别拿走按钮，返回对应卡片的索引。
-
-        拿走按钮位于卡片 box 范围内，通过判断图标命中 x 坐标是否落入各卡片
-        x_border 区间（左闭右闭）来确定归属卡片。
-        识别失败时返回第一张卡片的索引 作为兜底。
-
-        Args:
-            borders(list): 可选潜能卡片的边界框，每个元素为一个列表，包含2个元素，分别是左闭右闭的x轴边界
-            image(numpy.ndarray): 截图
-            max_try(int): 可选的最大重试次数，默认为1
-
-        Returns:
-            int: 目标卡片索引，识别失败时返回0。这里的索引是0-based的，适合给list使用。
-        """
-        result_boxes = self._template(
-            "星塔_节点_选择潜能_识别预选潜能位置_agent",
-            [],
-            image=image,
-            max_try=max_try
-        )
-
-        hit_x = result_boxes[0][0] if result_boxes else -1
-        matched = next(i for i, (low, high) in enumerate(borders) if low <= hit_x <= high)
-
-        if matched is None:
-            logger.error("拿走按钮识别失败，潜能选择可能会出现问题")
-            matched = 0
-
-        return matched
-
-    def match_trekker(self, trekker_image: numpy.ndarray, roi: list[int]) -> bool:
-        """匹配旅人"""
-        self.context.override_image("trekker_image", trekker_image)
-        reco_results = self._template(
-            "星塔_节点_选择潜能_识别旅人_agent",
-            [],
-            template="trekker_image",
-            roi=roi
-        )
-        return len(reco_results) != 0
-
-
-# endregion 使用maa与屏幕交互类 ==============================================================
-
-# region 潜能选择核心逻辑类 ===============================================================
-
-class ChoosePotentialHandler:
-    def __init__(self, screen: ScreenDataProcessor, data: Data):
-        self.screen = screen
-        self.data = data
-
-    def _wait_for_item_list_gone(self):
-        self.screen.load_last_screenshot()
-        while True:
-            if self.screen.check_item_list_visibility():
-                logger.debug("识别到干扰文字，等待1秒")
-                time.sleep(1)
-                self.screen.screenshot()
-                continue
-            break
-
-    def initialize_potentials(self):
-        # 初始化Potential对象，并储存到list中
-        potential_layouts = self.data.params.potential_layouts[self.data.potential_count]
-        potentials = [Potential(potential_layouts[i]) for i in range(self.data.potential_count)]
-
-        # 给潜能的selected、core字段赋值
-        self.data.selected_potential_index = self.screen.get_selected_potential_index(self.data.x_borders)
-        for i, p in enumerate(potentials):
-            p.index = i
-            if i == self.data.selected_potential_index:
-                p.selected = True
-            if self.data.core_potential:
-                p.core = True
-                p.old_level = 0
-                p.new_level = 1
-
-        return potentials
-
-    def read_potentials_info(self) -> Self:
-        """最原始的潜能信息识别器，仅识别推荐图标"""
-        self._wait_for_item_list_gone()
-        self.screen.screenshot()
-        self.data.potentials = self.initialize_potentials()
-
-        self._update_recommended_potentials()
-
-        return self
-
-    def choose(self) -> Potential | None:
-        """最原始的潜能选择，仅靠推荐图标选择潜能"""
-        potential = next((p for p in self.data.potentials if p.recommended), None)
-        if potential:
-            logger.info(f"[潜能选择] 推荐潜能")
-        return potential
-
-    def pick(self, potential: Potential) -> bool:
-        """点击潜能卡片"""
-        if potential.selected:
-            return True
-        return self.screen.click_potential(potential.box)
-
-    def dummy_potential(self, **kwargs) -> Potential:
-        """返回一个虚拟的潜能对象"""
-        potential_layout = self.data.params.potential_layouts[1][0]
-        base_potential = Potential(layout=potential_layout)
-        return replace(base_potential, **kwargs)
-
-    def _update_names(self):
-        rois = self.data.core_potential_name_rois if self.data.core_potential else self.data.general_potential_name_rois
-        adjusted_rois = self._get_adjusted_rois(rois)
-
-        for i, roi in enumerate(adjusted_rois):
-            self.data.potentials[i].name = self.screen.get_potential_name(roi)
-
-    def _update_levels(self):
-        if self.data.core_potential:
-            return
-
-        adjusted_rois = self._get_adjusted_rois(self.data.general_potential_level_rois)
-
-        for i, roi in enumerate(adjusted_rois):
-            old, new = self.screen.get_potential_level(roi)
-            self.data.potentials[i].old_level, self.data.potentials[i].new_level = old, new
-
-            # TODO: 测试用
-            if old == -1 and new == -1:
-                from utils.image_handler import save_image
-                save_image(self.screen.image, f"第{i}个潜能等级识别失败_{roi}")
-
-    def _update_recommended_potentials(self):
-        adjusted_rois = self._get_adjusted_rois(self.data.recommended_level_rois)
-        indices = self.screen.get_recommended_potential(self.data.x_borders)
-        for index in indices:
-            roi = adjusted_rois[index]
-            self.data.potentials[index].recommended = True
-            if self.data.core_potential:
-                self.data.potentials[index].recommended_level = 1
-            else:
-                self.data.potentials[index].recommended_level = self.screen.get_recommend_level(roi)
-
-    def _update_trekkers(self):
-        save_rois = self._get_adjusted_rois(self.data.trekker_rois)
-        expanded_rois = self._expand_rois(save_rois)
-        # 先根据State.trekker_images识别旅人，根据识别到的信息返回index给Potential的trekker字段
-        for potential_i, roi in enumerate(expanded_rois):
-            for trekker_i, trekker_image in enumerate(State.trekker_images):
-                if self.screen.match_trekker(trekker_image, roi):
-                    self.data.potentials[potential_i].trekker = str(trekker_i)
-                    break
-
-            # 识别不到时，截图保存到State中，并返回新的index给Potential的trekker字段
-            if not self.data.potentials[potential_i].trekker:
-                cropped_image = self.screen.crop_screenshot(save_rois[potential_i])
-                State.trekker_images.append(cropped_image)
-                self.data.potentials[potential_i].trekker = str(len(State.trekker_images) - 1)
-
-            # 如果trekker_images超过3个，输出错误日志
-            if len(State.trekker_images) > 3:
-                logger.error("识别到超过3种旅人，后续潜能判断将会出现问题")
-
-    def _get_adjusted_rois(self, base_rois: list[list[int]]) -> list[list[int]]:
-        """安全地获取偏移后的 ROI 副本，避免污染原始数据"""
-        selected_index = self.data.selected_potential_index
-        offset = self.data.params.selected_potential_offset
-
-        # 使用列表推导式创建副本并应用偏移
-        return [
-            [r[0], max(0, r[1] - offset), r[2], r[3]] if i == selected_index else list(r)
-            for i, r in enumerate(base_rois)
-        ]
-
-    @staticmethod
-    def _expand_rois(base_rois: list[list[int]], px: int = 15) -> list[list[int]]:
-        """扩展ROI"""
-        return [
-            [r[0] - px, r[1] - px, r[2] + px * 2, r[3] + px * 2]
-            for r in base_rois
-        ]
-
-    def choose_fallback_potential(self):
-        priority_rules = [
-            lambda p: p.recommended, # 1. 尝试取系统推荐潜能
-            lambda p: p.old_level > 0 # 2. 没有系统推荐潜能，选择之前抓过的潜能
-        ]
-
-        for rule in priority_rules:
-            candidates = [p for p in self.data.potentials if rule(p)]
-            if candidates:
-                # 按照等级跨度降序、推荐等级降序、旧等级降序来排序，选择最优的潜能
-                return max(candidates, key=lambda p: (p.level_span, p.recommended_level, p.old_level), default=None)
-
-        # 都没有的话，放弃选择，系统选了哪张就哪张
-        return self._default_potential
-
-    def refresh(self):
-        self.screen.refresh()
-        self.data.refresh_count += 1
-
-    @property
-    def _default_potential(self):
-        potential = next(p for p in self.data.potentials if p.selected)
-        return potential
-
-
-# endregion 潜能选择核心逻辑类 ===============================================================
-
-# region 潜能选择子类：系统推荐潜能+选择策略 ===============================================================
-
-class GameRecommendedHandler(ChoosePotentialHandler):
-    def __init__(self, screen: ScreenDataProcessor, data: Data):
-        super().__init__(screen, data)
-
-    def read_potentials_info(self) -> Self:
-        self._wait_for_item_list_gone()
-        self.screen.screenshot()
-        self.data.potentials = self.initialize_potentials()
-
-        self._update_recommended_potentials()
-        self._update_names()
-        self._update_levels()
-        self._update_trekkers()
-
-        # 输出当前潜能列表到日志
-        for potential in self.data.potentials:
-            recommended_output = f"系统推荐{potential.recommended_level}级" if potential.recommended_level > 0 else "无"
-            if self.data.core_potential:
-                logger.info(f"[潜能识别] {potential.name} | 核心潜能 | {recommended_output}")
-            else:
-                old = potential.old_level
-                new = potential.new_level
-                logger.info(f"[潜能识别] {potential.name} | 等级 {old}→{new} | {recommended_output}")
-
-            # TODO: 测试用
-            logger.debug(f"所属旅人 {potential.trekker} | 潜能数 {State.owned_potentials.count(trekker=potential.trekker)}")
-        if not self.data.core_potential:
-            p_counts = [State.owned_potentials.count(trekker=t) for t in ["0", "1", "2"]]
-            p_leveling_counts = [State.owned_potentials.count(trekker=t, leveling_only=True) for t in ["0", "1", "2"]]
-            logger.debug(f"当前分别持有潜能 {p_counts}，未满级潜能 {p_leveling_counts}")
-
-        return self
-
-    def choose(self) -> Potential | None:
-        # 根据参数选择潜能选择器
-        if self.data.params.chooser.startswith("tower_8"):
-            best_potential = self.tower_8_chooser()
-        else:
-            best_potential = self.choose_fallback_potential()
-
-        if best_potential:
-            logger.info(f"[潜能选择] {best_potential.name}")
-
-        return best_potential
-
-    def tower_8_chooser(self) -> Potential | None:
-        """
-        塔8专用策略，采用打分制
-        根据用户设置的最大刷新次数以及刷新分数阈值，判断是否需要刷新
-
-        Returns:
-            Potential | None: 最好的系统推荐潜能，若没有则返回 None
-        """
-
-        # 核心潜能选择，直接选择推荐潜能，如没有则选择默认潜能
-        if self.data.core_potential:
-            best_potential = max(
-                (p for p in self.data.potentials if p.recommended),
-                key=lambda p: p.recommended_level,
-                default=self._default_potential,
-            )
-            self._tower_8_record(best_potential)
-            return best_potential
-
-        # 打分
-        for p in self.data.potentials:
-            p.score = self._tower_8_score(p)
-
-        # 取得最优潜能
-        best_potential = max(
-            self.data.potentials,
-            key=lambda p: (
-                p.score,
-                p.level_span,
-                p.recommended_level,
-                p.old_level,
-            ),
-        )
-
-        # 不能刷：强化潜能、金币不足、刷新次数用尽等，直接选当前最优。
-        if not self.data.refresh_botton or not self.data.refreshable:
-            self._tower_8_record(best_potential)
-            return best_potential
-
-        # 当前牌到达刷新分数阈值，直接选当前最优，否则返回 None 让外层刷新。
-        score_threshold = self._tower_8_threshold()
-        logger.debug(f"当前最优牌 {best_potential.name} 得分 {best_potential.score}，阈值 {score_threshold}")
-        if best_potential.score >= score_threshold:
-            self._tower_8_record(best_potential)
-            return best_potential
-
-        return None
-
-    def _tower_8_score(self, p: Potential | OwnedPotential, *, level_span: int = 0) -> int | float:
-        """
-        塔8专用潜能打分核心函数，满分为100分
-        由于还没摸透潜能抽取规则，所以只能使用平均概率的方法去计算
-        默认目标为6+5+5=16个潜能
-        """
-        # 已获得潜能可能需要从保存的State类中获得推荐等级
-        if isinstance(p, Potential) and p.old_level > 0 and not p.recommended:
-            p.recommended_level = State.owned_potentials.find_recommended_level(p.name, trekker=p.trekker, fuzzy=True)
-            p.recommended = True if p.recommended_level >= 0 else False
-        # 传入OwnedPotential对象时，需要设置等级跨度，然后转为Potential对象处理
-        if isinstance(p, OwnedPotential):
-            if level_span == 0:
-                logger.warning("传入OwnedPotential对象时未设置等级跨度，将默认为1")
-                level_span = 1
-            new_level = p.level + level_span
-            recommended = True if p.recommended_level > 0 else False
-            p = self.dummy_potential(
-                name=p.name, trekker=p.trekker, old_level=p.level, new_level=new_level, recommended=recommended,
-                recommended_level=p.recommended_level, core=p.core
-            )
-
-        # 取得有效升级量，只有在升级到推荐等级上面才是有效升级
-        effective_gain = max(0, min(p.new_level, p.recommended_level) - min(p.old_level, p.recommended_level))
-
-        if p.old_level == 0:
-            # 新潜能
-            # 在辉光的奇迹buff没用完的情况下，新潜能升级量为3是100分，2是66分，1是33分，0是0分
-            # 在辉光的奇迹buff用完的情况下，新潜能升级量为2是100分，1是50分，0是0分
-            max_recommended_level = 3 if State.high_level_span_count < 10 else 2
-            score = effective_gain / max_recommended_level * 100
-        else:
-            # 老潜能，通过已保有潜能种类数打分
-            score = min(16, State.owned_potentials.count()) / 16 * 100
-            # 根据已有等级，对分数进行修正
-            score = score * (p.new_level / 6 * effective_gain)
-            # TODO: 如果所有潜能都获得了且满足推荐等级，则需要把非推荐潜能的分数提高
-            # TODO: 辉光的奇迹用在推荐等级2以下的潜能是否减分
-
-        return round(score, 2)
-
-    def _tower_8_threshold(self) -> int | float:
-        """
-        塔8专用刷新阈值计算函数
-        由于还没摸透潜能抽取规则，所以暂时只能使用平均概率的方法去计算
-        然后通过用户设置的激进程度去调整
-        默认目标为6+5+5=16个6级潜能
-        """
-        # 1. 统计新潜能的种类数
-        trekkers = ["0", "1", "2"]
-        owned_stats = {
-            t: {
-                "total": State.owned_potentials.count(trekker=t),
-                "leveling": State.owned_potentials.count(trekker=t, leveling_only=True)
-            }
-            for t in trekkers
-        }
-        new_potential_counts = [
-            0 if stats["total"] >= 5 and stats["leveling"] >= 3
-            else (12 - stats["total"])
-            for stats in owned_stats.values()
-        ]
-        new_potential_count = sum(new_potential_counts)
-
-        if new_potential_count == 0:
-            return 0
-
-        # 2. 计算新旧潜能的期望分数
-        new_potential_scores = self._calculate_new_potential_scores(new_potential_count)
-        old_potential_scores = self._calculate_old_potential_scores()
-        potential_scores = new_potential_scores + old_potential_scores
-
-        # 3. 计算特定潜能在3次抽取中成为最佳潜能的期望分数
-        if len(potential_scores) <= 3: # 卡池只剩3个潜能代表当前显示的3个跟刷新后的3个会一模一样，不需要刷新
-            return 0
-        potential_scores.sort()
-        potential_count = len(potential_scores)
-        expected_score = sum(
-            potential_scores[i] * comb(i, 2) / comb(potential_count, 3)
-            for i in range(2, potential_count)
-        )
-
-        # 4. 把刷新价值量化为潜能抽取的期望分数
-        refresh_cost_prop = self.data.refresh_cost / 160
-        refresh_cost_score = expected_score * refresh_cost_prop
-
-        # 5. 通过用户设置的刷新次数压力，调整刷新分数阈值
-        pressure = 1
-        threshold = expected_score - refresh_cost_score - (pressure * self.data.refresh_count)
-
-        return round(threshold, 2)
-
-    def _calculate_new_potential_scores(self, new_potential_count: int) -> list[int | float]:
-        """计算所有新潜能的总期望分数"""
-        # 新潜能等级概率权重: Lv1(80%), Lv2(10%), Lv3(10%)
-        level_weights = [(1, 0.8), (2, 0.1), (3, 0.1)] if State.high_level_span_count < 10 else [(1, 0.8), (2, 0.2)]
-
-        recommended_score = sum(
-            self._tower_8_score(self.dummy_potential(old_level=0, new_level=lv, recommended_level=6)) * weight
-            for lv, weight in level_weights
-        )
-        # 由于6/5的硬上限限制，所以即使拿到了垃圾潜能，也只能当作拿到推荐潜能算
-        new_recommended_potential_count = max(0, 16 - State.owned_potentials.count())
-        # 计算期望分数
-        recommended_scores = [recommended_score for _ in range(new_recommended_potential_count)]
-        unrecommended_scores = [0 for _ in range(new_potential_count - new_recommended_potential_count)]
-
-        return recommended_scores + unrecommended_scores
-
-    def _calculate_old_potential_scores(self) -> list[int | float]:
-        """计算所有已有潜能升一级的总分数"""
-        return [
-            self._tower_8_score(p, level_span=1)
-            for p in State.owned_potentials
-            if not p.core and p.level < 6
-        ]
-
-    @staticmethod
-    def _tower_8_record(p: Potential) -> None:
-        State.potentials_level_count += p.level_span
-        logger.info(f"潜能计数 {State.potentials_level_count}")
-        if not p.core:
-            if p.old_level == 0 and p.level_span >= 3:
-                State.high_level_span_count += 1
-                logger.info(f"辉光的奇迹计数 {State.high_level_span_count}/10")
-
-            if p.old_level > 0 and p.level_span == 2:
-                State.enhance_high_level_span_count += 1
-                logger.info(f"潜能飞升计数 {State.enhance_high_level_span_count}/5")
-
-
-# endregion 潜能选择子类：系统推荐潜能+选择策略 ===============================================================
-
-# region 潜能选择子类：系统推荐潜能+扫描背包+选择策略 ===============================================================
-
-class RecommendationPlusBagScanHandler(GameRecommendedHandler):
-    def __init__(self, screen: ScreenDataProcessor, data: Data):
-        super().__init__(screen, data)
-
-    # def _tower_8_score(self, p: Potential) -> int:
-    # 因为能获得整体情况，所以还需要考虑下面问题
-    # 还得考虑潜能到达旅人上限的问题，以保证推荐潜能数超过普通上限的情况打分偏向于能够突破上限
-    # 推荐等级越高分数越高的修正
-
-    # def _tower_8_threshold(self, threshold: int) -> int:
-    # 这个可能不用调整思路，但需要把默认16种潜能改为整个潜能库的概率计算
-
-
-# endregion 潜能选择子类：系统推荐潜能+扫描背包+选择策略 ===============================================================
-
-# region 潜能选择子类：json优先级选择 ===============================================================
-
-class AssistantPriorityHandler(ChoosePotentialHandler):
-    def __init__(self, screen: ScreenDataProcessor, data: Data):
-        super().__init__(screen, data)
-
-    def read_potentials_info(self) -> Self:
-        self._wait_for_item_list_gone()
-        self.screen.screenshot()
-        self.data.potentials = self.initialize_potentials()
-
-        self._update_names()
-        self._update_levels()
-
-        return self
-
-    def choose(self):
-        # 获得所有潜能的排名
-        self._update_priority()
-
-        # 输出比较结果
-        for potential in self.data.potentials:
-            print_rank = potential.rank + 1 if potential.rank >= 0 else "无"
-            if self.data.core_potential:
-                logger.info(f"[潜能识别] {potential.name} | 核心潜能 | 排名 {print_rank}")
-            else:
-                old = potential.old_level
-                new = potential.new_level
-                logger.info(f"[潜能识别] {potential.name} | 等级 {old}→{new} | 排名 {print_rank}")
-
-        # 选择排名最高的潜能
-        best_potential = self.best_potential
-        if best_potential:
-            logger.info(f"[潜能选择] {best_potential.name}")
-
-        return best_potential
-
-    def _update_priority(self):
-        for potential in self.data.potentials:
-            rank, sub_rank, trekker = self._get_potential_priority(potential)
-            potential.rank = rank
-            potential.sub_rank = sub_rank
-            potential.trekker = trekker
-
-    def _get_potential_priority(
-        self,
-        potential: Potential,
-    ) -> tuple[int, int, str | None]:
-        """获取单个待选潜能在规则列表中的最高排名及其 trekker 归属。
-
-        遍历 priority_list，找到所有名称匹配且满足 level_span / max_level / refresh
-        条件的规则，返回排名数值最小（即优先级最高）的规则对应的排名与 trekker。
-
-        Args:
-            potential: 单个待选潜能，结构：
-                {"name": str, "old_level": int, "new_level": int, "box": list}
-
-        Returns:
-            tuple[int, int, str | None]: (rank, sub_rank, trekker)
-                rank 为匹配到的最小排名数值； rank 从 0 开始。无匹配时返回 -1
-                sub_rank 为命中的 potential 名称在该规则 names 列表中的下标；无匹配时返回 -1
-                trekker 为对应规则的归属角色；无匹配时返回空字符
-        """
-        priority_list = self.data.params.priority_list
-
-        best_entry = None
-        best_rank = -1
-        best_sub_rank = -1
-
-        for rank, entry in enumerate(priority_list):
-            # 1. 基础剪枝：优先级如果不更高，直接跳过
-            if best_entry and entry["priority"] >= best_entry["priority"]:
-                continue
-
-            # 2. 匹配名称并获取优先级排名
-            sub_rank = self._find_sub_rank(potential.name, entry["names"])
-            if sub_rank == -1:
-                continue
-
-            # 3. 验证其他规则是否通过
-            if not self._is_entry_valid(entry, potential):
-                continue
-
-            # 全部通过后，记录该行及副等级
-            best_entry = entry
-            best_rank = rank
-            best_sub_rank = sub_rank
-
-        if not best_entry:
-            best_entry = {"trekker": ""}
-
-        return best_rank, best_sub_rank, best_entry["trekker"]
-
-    def _find_sub_rank(self, name: str, rule_names: list[str]) -> int:
-        """通过潜能名称获取最优排名数值"""
-        return next((i for i, r in enumerate(rule_names)
-                     if self._match_potential_name(name, r)), -1)
-
-    @staticmethod
-    def _match_potential_name(ocr_name: str, rule_name: str) -> bool:
-        """比较 OCR 识别的潜能名称与规则中的潜能名称是否匹配。
-
-        OCR 可能产生前后漏字或噪声字符，因此先清洗两边字符串（去除非 Unicode
-        字母数字字符），再用 in 检查，覆盖前后漏字的情况。
-
-        中间漏字或错字无法处理，属于 OCR 识别质量问题。
-
-        Args:
-            ocr_name: OCR 识别到的潜能名称
-            rule_name: 优先级规则中定义的潜能名称
-
-        Returns:
-            bool: 两者匹配时返回 True
-        """
-        cleaned_ocr = re.sub(r'\W', '', ocr_name)
-        cleaned_rule = re.sub(r'\W', '', rule_name)
-
-        # 防止"" in "任意字符串" 返回True
-        if not cleaned_ocr or not cleaned_rule:
-            return False
-
-        chars_to_remove = "ー"
-        table = str.maketrans("", "", chars_to_remove)
-        cleaned_ocr = cleaned_ocr.translate(table)
-        cleaned_rule = cleaned_rule.translate(table)
-
-        return cleaned_ocr in cleaned_rule
-
-    def _is_entry_valid(self, entry: dict, potential: Potential) -> bool:
-        """业务规则过滤器：方便未来随意扩展判定条件"""
-        # 核心潜能默认全部通过
-        if self.data.core_potential:
-            return True
-
-        # 普通潜能，组合匹配规则
-        checks = [
-            potential.old_level < entry["max_level"],
-            potential.level_span >= entry["level_span"],
-            self.data.refresh_count >= entry["refresh"]
-        ]
-        return all(checks)
-
-    @property
-    def best_potential(self) -> Potential | None:
-        """按照排名升序、等级跨度降序、副排名升序三个维度，筛选出最好的潜能"""
-        valid_potentials = (p for p in self.data.potentials if p.rank >= 0)
-        return min(valid_potentials, key=lambda p: (p.rank, -p.level_span, p.sub_rank), default=None)
-
-
-# endregion 潜能选择子类：json优先级选择 ===============================================================
-
-# region 主函数 ===============================================================
-
-@AgentServer.custom_recognition("choose_potential_recognition")
-class ChoosePotentialRecognition(CustomRecognition):
-
-    def analyze(
-            self,
-            context: Context,
-            argv: CustomRecognition.AnalyzeArg,
-    ) -> CustomRecognition.AnalyzeResult:
-        """自动选择潜能的主流程。
-
-        读取 attach 参数与已拥有潜能状态，识别当前三张潜能卡片，
-        根据自定义优先级列表选出最优潜能并返回其 box；
-        不满足条件时执行刷新，最终兜底识别系统推荐图标。
-        选择完成后将已拥有潜能状态写回节点 attach 持久化。
-
-        Args:
-            context: maa.context.Context
-            argv: CustomAction.RunArg，含当前截图与节点名
-
-        Returns:
-            bool: 返回 True
-        """
-        node_name = argv.node_name
-        params = self._get_params(context, node_name)
-        data = Data(params=params)
-        screen = ScreenDataProcessor(context)
-
-        # 获取只使用一次的数据
-        image = argv.image
-        data.current_coin = screen.get_current_coin(image)
-        data.refresh_cost = screen.get_refresh_cost(image)
-        data.core_potential = screen.check_core_potential(image)
-        data.potential_count = screen.get_potential_count(data.core_potential, image)
-
-        # 加载相应的潜能处理类
-        if data.params.handler == "json":
-            handler = AssistantPriorityHandler(screen, data)
-        elif data.params.handler == "default+":
-            handler = GameRecommendedHandler(screen, data)
-        elif data.params.handler == "default++":
-            handler = RecommendationPlusBagScanHandler(screen, data)
-        else: # default
-            handler = ChoosePotentialHandler(screen, data)
-
-        while True:
-            # 获取潜能数据，并选择潜能
-            potential = handler.read_potentials_info().choose()
-            if potential:
-                break
-            elif data.refreshable:
-                logger.info("没有找到符合条件的潜能，尝试刷新")
-                handler.refresh()
-            else:
-                logger.info("[潜能选择] 没有找到符合条件的潜能，将按照保底顺序选择")
-                potential = handler.choose_fallback_potential()
-                break
-
-        # 点击潜能
-        click_result = handler.pick(potential)
-        if not click_result:
-            logger.error(f"点击潜能失败")
-
-        # 保存已选潜能数据到状态类中
-        State.owned_potentials.save(
-            potential,
-            fuzzy=data.params.handler in {"default+", "default++"},
-        )
-
-        return CustomRecognition.AnalyzeResult(box=potential.box, detail={})
-
-    @staticmethod
-    def _get_params(context: Context, node_name: str) -> Parameters:
-        """获取节点 attach 中的所有参数，缺失时返回安全默认值。
-
-        Args:
-            context: maa.context.Context
-            node_name: 当前节点名称，用于动态获取节点数据
-
-        Returns:
-            Parameters: 包含以下属性的实例：
-                - max_refresh_count (int): 最大刷新次数，0 表示禁用
-                - reserved_coin (int): 预留金币，计算可用金币时需减去此值
-                - priority_list (list): 自定义优先级列表
-                - owned_potentials (dict): 已拥有潜能状态，按 trekker 分组
-        """
-        node_data = context.get_node_data(node_name)
-        attach = node_data.get("attach", {})
-        params = Parameters(**attach)
-        return params
-
-
-# endregion 主函数 ===============================================================
