+++ agent/custom/reco/climb_tower_potential/__init__.py
@@ -1,0 +1,7 @@
+from custom.reco.climb_tower_potential.main import *
+from custom.reco.climb_tower_potential.bag_scan import *
+
+__all__ = [
+    "ChoosePotentialRecognition",
+    "PotentialBagTestRecognition"
+]
